Tidy up `pw_hash` function
* Remove legacy non-namespaced `lib/puppet/parser/functions/apache_pw_hash.rb`
* Create 4x API non-namespaced `apache_pw_hash` shim that calls `apache::pw_hash`
* Rename existing 4x API `apache::apache_pw_hash` to `apache::pw_hash`
* Create shim for `apache::apache_pw_hash` that calls `apache::pw_hash`
* Mark shims as DEPRECATED
* Test all 3 versions work
