+++ spec/functions/pw_hash_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe 'apache::apache_pw_hash' do
+shared_examples 'apache::pw_hash function' do
   it { is_expected.not_to eq(nil) }
   it { is_expected.to run.with_params.and_raise_error(ArgumentError) }
   it { is_expected.to run.with_params('').and_raise_error(ArgumentError) }
@@ -9,4 +9,17 @@ describe 'apache::apache_pw_hash' do
   it { is_expected.to run.with_params({}).and_raise_error(ArgumentError) }
   it { is_expected.to run.with_params([]).and_raise_error(ArgumentError) }
   it { is_expected.to run.with_params('test').and_return('{SHA}qUqP5cyxm6YcTAhz05Hph5gvu9M=') }
+end
+
+describe 'apache::pw_hash' do
+  it_behaves_like 'apache::pw_hash function'
+
+  describe 'deprecated shims' do
+    describe 'apache_pw_hash', type: :puppet_function do
+      it_behaves_like 'apache::pw_hash function'
+    end
+    describe 'apache::apache_pw_hash', type: :puppet_function do
+      it_behaves_like 'apache::pw_hash function'
+    end
+  end
 end
